Make the DocumentUrlDB not store the entire url
Instead it uses the IdTreeDb and IDFilenameDb to only store the filename
along with the parent folder's doc id.

We need more tests for this since it isn't well tested, and the code is
decently complex with many edge cases. Plus, since the DocumentUrlDB now
stat's every part of the url, it's quite expensive and introduces a hard
dependency with the filesystem, which makes testing much harder.

+++ src/engine/idutils.h
@@ -34,7 +34,7 @@ namespace Baloo {
  */
 inline quint64 statBufToId(const QT_STATBUF& stBuf)
 {
-    quint64 arr[2];
+    quint32 arr[2];
     arr[0] = stBuf.st_dev;
     arr[1] = stBuf.st_ino;
